Back End Finished
Back end finished,
Even tho i will have to modify it in the future because i saved everything in the DataBase and the README said that i gotta do at least one petition on the front so.
On other comments the getDogsByID does not bring the associated temperaments so i gotta check that.